correction on row 49
Replaced the str.strip with str.replace. 

Running the script with the old strip method was faulty when target apk-file had any of the characters a, p or k in its name. str.replace is the correct way to strip the ".apk" out of the name, I believe. Works for me at least...